feat(csa-review): batch cumulative review every N commits (#770)
Adds [review].batch_commits config (default 1 = existing behavior).
When >=2, commit/dev2merge workflows skip their intermediate cumulative
csa review if branch has fewer than N new commits since the last passed
cumulative review on this branch. Pre-push hook still forces HEAD review
before push, so batching only suppresses intermediate rounds, not the
push gate.

Last-passed SHA tracked per-branch at .csa/state/review/last-cumulative-
<branch>.txt, written only when review exit=0 and severity_summary has
zero HIGH/CRITICAL findings. Override via CSA_REVIEW_NOW=1.

Closes #770.
